Reorganize Python modules (#978)
Changes:

    Moved frame.py and series.py to their own eager module
    Split up the code under the lazy module into four files: frame.py, expr.py, expr_functions.py, and whenthen.py
    Moved the single function from ffi.py to utils.py.
    Specified the __all__ dunder for all modules; specifying exactly which objects must be available at the module level. This means that things like numpy aren't available anymore through the polars module; e.g. you can no longer use polars.np.array(). And that's good! Only the relevant functions/classes are available at the top level (if you want to check exactly which ones, import polars and run polars.__dict__.keys())
    Removed all explicit relative imports of the 'problematic' modules that could cause cyclic imports, like Series and DataFrame. Instead, modules do import polars as pl and access the objects through pl.Series, pl.DataFrame, etc.. That fixes most cyclic import issues.
    Stringified most type hint references to the polars objects; it wasn't such a bad idea after all... it solves a few issues with imports, and still works as intended for mypy.
