+++ py-polars/polars/__init__.py
@@ -1,16 +1,13 @@
 # flake8: noqa
 from .datatypes import *
-from .frame import DataFrame, StringCache, wrap_df
+from .eager import *
 from .functions import *
 from .lazy import *
-from .series import Series, wrap_s
 
 # during docs building the binary code is not yet available
 try:
-    from .frame import version
+    from .polars import version
 
     __version__ = version()
 except ImportError:
     pass
-
-__pdoc__ = {"ffi": False}
