+++ py-polars/polars/utils.py
@@ -1,8 +1,16 @@
+import ctypes
 import typing as tp
 import warnings
-from typing import Dict, Tuple, Union
+from typing import Any, Dict, Tuple, Union
 
+import numpy as np
 import pyarrow as pa
+
+__all__ = [
+    "coerce_arrow",
+    "_process_null_values",
+    "_ptr_to_numpy",
+]
 
 
 def coerce_arrow(array: pa.Array) -> pa.Array:
@@ -47,3 +55,26 @@ def _process_null_values(
         return list(null_values.items())
     else:
         return null_values
+
+
+# https://stackoverflow.com/questions/4355524/getting-data-from-ctypes-array-into-numpy
+def _ptr_to_numpy(ptr: int, len: int, ptr_type: Any) -> np.ndarray:
+    """
+
+    Parameters
+    ----------
+    ptr
+        C/Rust ptr casted to usize.
+    len
+        Length of the array values.
+    ptr_type
+        Example:
+            f32: ctypes.c_float)
+
+    Returns
+    -------
+    View of memory block as numpy array.
+
+    """
+    ptr_ctype = ctypes.cast(ptr, ctypes.POINTER(ptr_type))
+    return np.ctypeslib.as_array(ptr_ctype, (len,))
